+++ website/app/themes/lcds/functions.php
@@ -19,5 +19,6 @@ require_once __DIR__ . '/inc/acf.php';
 require_once __DIR__ . '/inc/editor.php';
 require_once __DIR__ . '/inc/options.php';
 require_once __DIR__ . '/inc/roles.php';
+require_once __DIR__ . '/inc/logins.php';
 require_once __DIR__ . '/inc/seo.php';
 require_once __DIR__ . '/inc/contacts.php';